backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-CORS
Authlib
passlib[bcrypt]
python-dotenv
psycopg2-binary
requests
Werkzeug
gunicorn
firebase-admin